Split fat module
