+++ src/fat/volume.rs
@@ -1,14 +1,4 @@
-//! embedded-sdmmc-rs - FAT16/FAT32 file system implementation
-//!
-//! Implements the File Allocation Table file system. Supports FAT16 and FAT32 volumes.
-
-use crate::blockdevice::BlockCount;
-use crate::{
-    Attributes, Block, BlockDevice, BlockIdx, Cluster, Controller, DirEntry, Directory, Error,
-    ShortFileName, TimeSource, Timestamp, VolumeType,
-};
-use byteorder::{ByteOrder, LittleEndian};
-use core::convert::TryFrom;
+//! FAT volume
 
 #[cfg(feature = "log")]
 use log::{debug, trace, warn};
@@ -16,48 +6,16 @@ use log::{debug, trace, warn};
 #[cfg(feature = "defmt-log")]
 use defmt::{debug, trace, warn};
 
-/// Number of entries reserved at the start of a File Allocation Table
-pub const RESERVED_ENTRIES: u32 = 2;
-
-/// Indentifies the supported types of FAT format
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum FatType {
-    /// FAT16 Format
-    Fat16,
-    /// FAT32 Format
-    Fat32,
-}
-
-/// Indentifies the supported types of FAT format
-#[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
-#[derive(Debug, Eq, PartialEq)]
-pub enum FatSpecificInfo {
-    /// Fat16 Format
-    Fat16(Fat16Info),
-    /// Fat32 Format
-    Fat32(Fat32Info),
-}
-
-/// FAT32 specific data
-#[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
-#[derive(Debug, Eq, PartialEq)]
-pub struct Fat32Info {
-    /// The root directory does not have a reserved area in FAT32. This is the
-    /// cluster it starts in (nominally 2).
-    pub(crate) first_root_dir_cluster: Cluster,
-    /// Block idx of the info sector
-    pub(crate) info_location: BlockIdx,
-}
-
-/// FAT16 specific data
-#[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
-#[derive(Debug, Eq, PartialEq)]
-pub struct Fat16Info {
-    /// The block the root directory starts in. Relative to start of partition (so add `self.lba_offset` before passing to controller)
-    pub(crate) first_root_dir_block: BlockCount,
-    /// Number of entries in root directory (it's reserved and not in the FAT)
-    pub(crate) root_entries_count: u16,
-}
+use crate::{
+    fat::{
+        Bpb, Fat16Info, Fat32Info, FatSpecificInfo, FatType, InfoSector, OnDiskDirEntry,
+        RESERVED_ENTRIES,
+    },
+    Attributes, Block, BlockCount, BlockDevice, BlockIdx, Cluster, Controller, DirEntry, Directory,
+    Error, ShortFileName, TimeSource, VolumeType,
+};
+use byteorder::{ByteOrder, LittleEndian};
+use core::convert::TryFrom;
 
 /// The name given to a particular FAT formatted volume.
 #[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
@@ -70,6 +28,15 @@ impl VolumeName {
     /// Create a new VolumeName
     pub fn new(data: [u8; 11]) -> VolumeName {
         VolumeName { data }
+    }
+}
+
+impl core::fmt::Debug for VolumeName {
+    fn fmt(&self, fmt: &mut core::fmt::Formatter) -> core::fmt::Result {
+        match core::str::from_utf8(&self.data) {
+            Ok(s) => write!(fmt, "{:?}", s),
+            Err(_e) => write!(fmt, "{:?}", &self.data),
+        }
     }
 }
 
@@ -97,362 +64,6 @@ pub struct FatVolume {
     pub(crate) cluster_count: u32,
     /// Type of FAT
     pub(crate) fat_specific_info: FatSpecificInfo,
-}
-
-impl core::fmt::Debug for VolumeName {
-    fn fmt(&self, fmt: &mut core::fmt::Formatter) -> core::fmt::Result {
-        match core::str::from_utf8(&self.data) {
-            Ok(s) => write!(fmt, "{:?}", s),
-            Err(_e) => write!(fmt, "{:?}", &self.data),
-        }
-    }
-}
-
-/// Represents a Boot Parameter Block. This is the first sector of a FAT
-/// formatted partition, and it describes various properties of the FAT
-/// filesystem.
-pub struct Bpb<'a> {
-    data: &'a [u8; 512],
-    fat_type: FatType,
-    cluster_count: u32,
-}
-
-impl<'a> Bpb<'a> {
-    const FOOTER_VALUE: u16 = 0xAA55;
-
-    /// Attempt to parse a Boot Parameter Block from a 512 byte sector.
-    pub fn create_from_bytes(data: &[u8; 512]) -> Result<Bpb, &'static str> {
-        let mut bpb = Bpb {
-            data,
-            fat_type: FatType::Fat16,
-            cluster_count: 0,
-        };
-        if bpb.footer() != Self::FOOTER_VALUE {
-            return Err("Bad BPB footer");
-        }
-
-        let root_dir_blocks = ((u32::from(bpb.root_entries_count()) * OnDiskDirEntry::LEN_U32)
-            + (Block::LEN_U32 - 1))
-            / Block::LEN_U32;
-        let data_blocks = bpb.total_blocks()
-            - (u32::from(bpb.reserved_block_count())
-                + (u32::from(bpb.num_fats()) * bpb.fat_size())
-                + root_dir_blocks);
-        bpb.cluster_count = data_blocks / u32::from(bpb.blocks_per_cluster());
-        if bpb.cluster_count < 4085 {
-            return Err("FAT12 is unsupported");
-        } else if bpb.cluster_count < 65525 {
-            bpb.fat_type = FatType::Fat16;
-        } else {
-            bpb.fat_type = FatType::Fat32;
-        }
-
-        match bpb.fat_type {
-            FatType::Fat16 => Ok(bpb),
-            FatType::Fat32 if bpb.fs_ver() == 0 => {
-                // Only support FAT32 version 0.0
-                Ok(bpb)
-            }
-            _ => Err("Invalid FAT format"),
-        }
-    }
-
-    // FAT16/FAT32
-    define_field!(bytes_per_block, u16, 11);
-    define_field!(blocks_per_cluster, u8, 13);
-    define_field!(reserved_block_count, u16, 14);
-    define_field!(num_fats, u8, 16);
-    define_field!(root_entries_count, u16, 17);
-    define_field!(total_blocks16, u16, 19);
-    define_field!(media, u8, 21);
-    define_field!(fat_size16, u16, 22);
-    define_field!(blocks_per_track, u16, 24);
-    define_field!(num_heads, u16, 26);
-    define_field!(hidden_blocks, u32, 28);
-    define_field!(total_blocks32, u32, 32);
-    define_field!(footer, u16, 510);
-
-    // FAT32 only
-    define_field!(fat_size32, u32, 36);
-    define_field!(fs_ver, u16, 42);
-    define_field!(first_root_dir_cluster, u32, 44);
-    define_field!(fs_info, u16, 48);
-    define_field!(backup_boot_block, u16, 50);
-
-    /// Get the OEM name string for this volume
-    pub fn oem_name(&self) -> &[u8] {
-        &self.data[3..11]
-    }
-
-    // FAT16/FAT32 functions
-
-    /// Get the Volume Label string for this volume
-    pub fn volume_label(&self) -> &[u8] {
-        if self.fat_type != FatType::Fat32 {
-            &self.data[43..=53]
-        } else {
-            &self.data[71..=81]
-        }
-    }
-
-    // FAT32 only functions
-
-    /// On a FAT32 volume, return the free block count from the Info Block. On
-    /// a FAT16 volume, returns None.
-    pub fn fs_info_block(&self) -> Option<BlockCount> {
-        if self.fat_type != FatType::Fat32 {
-            None
-        } else {
-            Some(BlockCount(u32::from(self.fs_info())))
-        }
-    }
-
-    // Magic functions that get the right FAT16/FAT32 result
-
-    /// Get the size of the File Allocation Table in blocks.
-    pub fn fat_size(&self) -> u32 {
-        let result = u32::from(self.fat_size16());
-        if result != 0 {
-            result
-        } else {
-            self.fat_size32()
-        }
-    }
-
-    /// Get the total number of blocks in this filesystem.
-    pub fn total_blocks(&self) -> u32 {
-        let result = u32::from(self.total_blocks16());
-        if result != 0 {
-            result
-        } else {
-            self.total_blocks32()
-        }
-    }
-
-    /// Get the total number of clusters in this filesystem.
-    pub fn total_clusters(&self) -> u32 {
-        self.cluster_count
-    }
-}
-
-/// File System Information structure is only present on FAT32 partitions. It
-/// may contain a valid number of free clusters and the number of the next
-/// free cluster. The information contained in the structure must be
-/// considered as advisory only. File system driver implementations are not
-/// required to ensure that information within the structure is kept
-/// consistent.
-pub struct InfoSector<'a> {
-    data: &'a [u8; 512],
-}
-
-impl<'a> InfoSector<'a> {
-    const LEAD_SIG: u32 = 0x4161_5252;
-    const STRUC_SIG: u32 = 0x6141_7272;
-    const TRAIL_SIG: u32 = 0xAA55_0000;
-
-    /// Try and create a new Info Sector from a block.
-    pub fn create_from_bytes(data: &[u8; 512]) -> Result<InfoSector, &'static str> {
-        let info = InfoSector { data };
-        if info.lead_sig() != Self::LEAD_SIG {
-            return Err("Bad lead signature on InfoSector");
-        }
-        if info.struc_sig() != Self::STRUC_SIG {
-            return Err("Bad struc signature on InfoSector");
-        }
-        if info.trail_sig() != Self::TRAIL_SIG {
-            return Err("Bad trail signature on InfoSector");
-        }
-        Ok(info)
-    }
-
-    define_field!(lead_sig, u32, 0);
-    define_field!(struc_sig, u32, 484);
-    define_field!(free_count, u32, 488);
-    define_field!(next_free, u32, 492);
-    define_field!(trail_sig, u32, 508);
-
-    /// Return how many free clusters are left in this volume, if known.
-    pub fn free_clusters_count(&self) -> Option<u32> {
-        match self.free_count() {
-            0xFFFF_FFFF => None,
-            n => Some(n),
-        }
-    }
-
-    /// Return the number of the next free cluster, if known.
-    pub fn next_free_cluster(&self) -> Option<Cluster> {
-        match self.next_free() {
-            // 0 and 1 are reserved clusters
-            0xFFFF_FFFF | 0 | 1 => None,
-            n => Some(Cluster(n)),
-        }
-    }
-}
-
-/// Represents a 32-byte directory entry as stored on-disk in a directory file.
-pub struct OnDiskDirEntry<'a> {
-    data: &'a [u8],
-}
-
-impl<'a> core::fmt::Debug for OnDiskDirEntry<'a> {
-    fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
-        write!(f, "OnDiskDirEntry<")?;
-        write!(f, "raw_attr = {}", self.raw_attr())?;
-        write!(f, ", create_time = {}", self.create_time())?;
-        write!(f, ", create_date = {}", self.create_date())?;
-        write!(f, ", last_access_data = {}", self.last_access_data())?;
-        write!(f, ", first_cluster_hi = {}", self.first_cluster_hi())?;
-        write!(f, ", write_time = {}", self.write_time())?;
-        write!(f, ", write_date = {}", self.write_date())?;
-        write!(f, ", first_cluster_lo = {}", self.first_cluster_lo())?;
-        write!(f, ", file_size = {}", self.file_size())?;
-        write!(f, ", is_end = {}", self.is_end())?;
-        write!(f, ", is_valid = {}", self.is_valid())?;
-        write!(f, ", is_lfn = {}", self.is_lfn())?;
-        write!(
-            f,
-            ", first_cluster_fat32 = {:?}",
-            self.first_cluster_fat32()
-        )?;
-        write!(
-            f,
-            ", first_cluster_fat16 = {:?}",
-            self.first_cluster_fat16()
-        )?;
-        write!(f, ">")?;
-        Ok(())
-    }
-}
-
-/// Represents the 32 byte directory entry. This is the same for FAT16 and
-/// FAT32 (except FAT16 doesn't use first_cluster_hi).
-impl<'a> OnDiskDirEntry<'a> {
-    pub(crate) const LEN: usize = 32;
-    pub(crate) const LEN_U32: u32 = 32;
-
-    define_field!(raw_attr, u8, 11);
-    define_field!(create_time, u16, 14);
-    define_field!(create_date, u16, 16);
-    define_field!(last_access_data, u16, 18);
-    define_field!(first_cluster_hi, u16, 20);
-    define_field!(write_time, u16, 22);
-    define_field!(write_date, u16, 24);
-    define_field!(first_cluster_lo, u16, 26);
-    define_field!(file_size, u32, 28);
-
-    /// Create a new on-disk directory entry from a block of 32 bytes read
-    /// from a directory file.
-    pub fn new(data: &[u8]) -> OnDiskDirEntry {
-        OnDiskDirEntry { data }
-    }
-
-    /// Is this the last entry in the directory?
-    pub fn is_end(&self) -> bool {
-        self.data[0] == 0x00
-    }
-
-    /// Is this a valid entry?
-    pub fn is_valid(&self) -> bool {
-        !self.is_end() && (self.data[0] != 0xE5)
-    }
-
-    /// Is this a Long Filename entry?
-    pub fn is_lfn(&self) -> bool {
-        let attributes = Attributes::create_from_fat(self.raw_attr());
-        attributes.is_lfn()
-    }
-
-    /// If this is an LFN, get the contents so we can re-assemble the filename.
-    pub fn lfn_contents(&self) -> Option<(bool, u8, [char; 13])> {
-        if self.is_lfn() {
-            let mut buffer = [' '; 13];
-            let is_start = (self.data[0] & 0x40) != 0;
-            let sequence = self.data[0] & 0x1F;
-            // LFNs store UCS-2, so we can map from 16-bit char to 32-bit char without problem.
-            buffer[0] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[1..=2]))).unwrap();
-            buffer[1] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[3..=4]))).unwrap();
-            buffer[2] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[5..=6]))).unwrap();
-            buffer[3] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[7..=8]))).unwrap();
-            buffer[4] = core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[9..=10])))
-                .unwrap();
-            buffer[5] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[14..=15])))
-                    .unwrap();
-            buffer[6] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[16..=17])))
-                    .unwrap();
-            buffer[7] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[18..=19])))
-                    .unwrap();
-            buffer[8] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[20..=21])))
-                    .unwrap();
-            buffer[9] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[22..=23])))
-                    .unwrap();
-            buffer[10] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[24..=25])))
-                    .unwrap();
-            buffer[11] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[28..=29])))
-                    .unwrap();
-            buffer[12] =
-                core::char::from_u32(u32::from(LittleEndian::read_u16(&self.data[30..=31])))
-                    .unwrap();
-            Some((is_start, sequence, buffer))
-        } else {
-            None
-        }
-    }
-
-    /// Does this on-disk entry match the given filename?
-    pub fn matches(&self, sfn: &ShortFileName) -> bool {
-        self.data[0..11] == sfn.contents
-    }
-
-    /// Which cluster, if any, does this file start at? Assumes this is from a FAT32 volume.
-    pub fn first_cluster_fat32(&self) -> Cluster {
-        let cluster_no =
-            (u32::from(self.first_cluster_hi()) << 16) | u32::from(self.first_cluster_lo());
-        Cluster(cluster_no)
-    }
-
-    /// Which cluster, if any, does this file start at? Assumes this is from a FAT16 volume.
-    fn first_cluster_fat16(&self) -> Cluster {
-        let cluster_no = u32::from(self.first_cluster_lo());
-        Cluster(cluster_no)
-    }
-
-    /// Convert the on-disk format into a DirEntry
-    pub fn get_entry(
-        &self,
-        fat_type: FatType,
-        entry_block: BlockIdx,
-        entry_offset: u32,
-    ) -> DirEntry {
-        let mut result = DirEntry {
-            name: ShortFileName {
-                contents: [0u8; 11],
-            },
-            mtime: Timestamp::from_fat(self.write_date(), self.write_time()),
-            ctime: Timestamp::from_fat(self.create_date(), self.create_time()),
-            attributes: Attributes::create_from_fat(self.raw_attr()),
-            cluster: if fat_type == FatType::Fat32 {
-                self.first_cluster_fat32()
-            } else {
-                self.first_cluster_fat16()
-            },
-            size: self.file_size(),
-            entry_block,
-            entry_offset,
-        };
-        result.name.contents.copy_from_slice(&self.data[0..11]);
-        result
-    }
 }
 
 impl FatVolume {
@@ -1435,301 +1046,3 @@ where
         }
     }
 }
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    fn parse(input: &str) -> Vec<u8> {
-        let mut output = Vec::new();
-        for line in input.lines() {
-            let line = line.trim();
-            if !line.is_empty() {
-                // 32 bytes per line
-                for index in 0..32 {
-                    let start = index * 2;
-                    let end = start + 1;
-                    let piece = &line[start..=end];
-                    let value = u8::from_str_radix(piece, 16).unwrap();
-                    output.push(value);
-                }
-            }
-        }
-        output
-    }
-
-    /// This is the first block of this directory listing.
-    /// total 19880
-    /// -rw-r--r-- 1 jonathan jonathan   10841 2016-03-01 19:56:36.000000000 +0000  bcm2708-rpi-b.dtb
-    /// -rw-r--r-- 1 jonathan jonathan   11120 2016-03-01 19:56:34.000000000 +0000  bcm2708-rpi-b-plus.dtb
-    /// -rw-r--r-- 1 jonathan jonathan   10871 2016-03-01 19:56:36.000000000 +0000  bcm2708-rpi-cm.dtb
-    /// -rw-r--r-- 1 jonathan jonathan   12108 2016-03-01 19:56:36.000000000 +0000  bcm2709-rpi-2-b.dtb
-    /// -rw-r--r-- 1 jonathan jonathan   12575 2016-03-01 19:56:36.000000000 +0000  bcm2710-rpi-3-b.dtb
-    /// -rw-r--r-- 1 jonathan jonathan   17920 2016-03-01 19:56:38.000000000 +0000  bootcode.bin
-    /// -rw-r--r-- 1 jonathan jonathan     136 2015-11-21 20:28:30.000000000 +0000  cmdline.txt
-    /// -rw-r--r-- 1 jonathan jonathan    1635 2015-11-21 20:28:30.000000000 +0000  config.txt
-    /// -rw-r--r-- 1 jonathan jonathan   18693 2016-03-01 19:56:30.000000000 +0000  COPYING.linux
-    /// -rw-r--r-- 1 jonathan jonathan    2505 2016-03-01 19:56:38.000000000 +0000  fixup_cd.dat
-    /// -rw-r--r-- 1 jonathan jonathan    6481 2016-03-01 19:56:38.000000000 +0000  fixup.dat
-    /// -rw-r--r-- 1 jonathan jonathan    9722 2016-03-01 19:56:38.000000000 +0000  fixup_db.dat
-    /// -rw-r--r-- 1 jonathan jonathan    9724 2016-03-01 19:56:38.000000000 +0000  fixup_x.dat
-    /// -rw-r--r-- 1 jonathan jonathan     110 2015-11-21 21:32:06.000000000 +0000  issue.txt
-    /// -rw-r--r-- 1 jonathan jonathan 4046732 2016-03-01 19:56:40.000000000 +0000  kernel7.img
-    /// -rw-r--r-- 1 jonathan jonathan 3963140 2016-03-01 19:56:38.000000000 +0000  kernel.img
-    /// -rw-r--r-- 1 jonathan jonathan    1494 2016-03-01 19:56:34.000000000 +0000  LICENCE.broadcom
-    /// -rw-r--r-- 1 jonathan jonathan   18974 2015-11-21 21:32:06.000000000 +0000  LICENSE.oracle
-    /// drwxr-xr-x 2 jonathan jonathan    8192 2016-03-01 19:56:54.000000000 +0000  overlays
-    /// -rw-r--r-- 1 jonathan jonathan  612472 2016-03-01 19:56:40.000000000 +0000  start_cd.elf
-    /// -rw-r--r-- 1 jonathan jonathan 4888200 2016-03-01 19:56:42.000000000 +0000  start_db.elf
-    /// -rw-r--r-- 1 jonathan jonathan 2739672 2016-03-01 19:56:40.000000000 +0000  start.elf
-    /// -rw-r--r-- 1 jonathan jonathan 3840328 2016-03-01 19:56:44.000000000 +0000  start_x.elf
-    /// drwxr-xr-x 2 jonathan jonathan    8192 2015-12-05 21:55:06.000000000 +0000 'System Volume Information'
-    #[test]
-    fn test_dir_entries() {
-        #[derive(Debug)]
-        enum Expected {
-            Lfn(bool, u8, [char; 13]),
-            Short(DirEntry),
-        }
-        let raw_data = r#"
-        626f6f7420202020202020080000699c754775470000699c7547000000000000 boot       ...i.uGuG..i.uG......
-        416f007600650072006c000f00476100790073000000ffffffff0000ffffffff Ao.v.e.r.l...Ga.y.s.............
-        4f5645524c4159532020201000001b9f6148614800001b9f6148030000000000 OVERLAYS   .....aHaH....aH......
-        422d0070006c00750073000f00792e006400740062000000ffff0000ffffffff B-.p.l.u.s...y..d.t.b...........
-        01620063006d00320037000f0079300038002d0072007000690000002d006200 .b.c.m.2.7...y0.8.-.r.p.i...-.b.
-        42434d3237307e31445442200064119f614861480000119f61480900702b0000 BCM270~1DTB .d..aHaH....aH..p+..
-        4143004f005000590049000f00124e0047002e006c0069006e00000075007800 AC.O.P.Y.I....N.G...l.i.n...u.x.
-        434f5059494e7e314c494e2000000f9f6148614800000f9f6148050005490000 COPYIN~1LIN ....aHaH....aH...I..
-        4263006f006d000000ffff0f0067ffffffffffffffffffffffff0000ffffffff Bc.o.m.......g..................
-        014c004900430045004e000f0067430045002e00620072006f00000061006400 .L.I.C.E.N...gC.E...b.r.o...a.d.
-        4c4943454e437e3142524f200000119f614861480000119f61480800d6050000 LICENC~1BRO ....aHaH....aH......
-        422d0062002e00640074000f001962000000ffffffffffffffff0000ffffffff B-.b...d.t....b.................
-        01620063006d00320037000f0019300039002d0072007000690000002d003200 .b.c.m.2.7....0.9.-.r.p.i...-.2.
-        42434d3237307e34445442200064129f614861480000129f61480f004c2f0000 BCM270~4DTB .d..aHaH....aH..L/..
-        422e0064007400620000000f0059ffffffffffffffffffffffff0000ffffffff B..d.t.b.....Y..................
-        01620063006d00320037000f0059300038002d0072007000690000002d006200 .b.c.m.2.7...Y0.8.-.r.p.i...-.b.
-        "#;
-        let results = [
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str_mixed_case("boot").unwrap(),
-                mtime: Timestamp::from_calendar(2015, 11, 21, 19, 35, 18).unwrap(),
-                ctime: Timestamp::from_calendar(2015, 11, 21, 19, 35, 18).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::VOLUME),
-                cluster: Cluster(0),
-                size: 0,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                1,
-                [
-                    'o', 'v', 'e', 'r', 'l', 'a', 'y', 's', '\u{0000}', '\u{ffff}', '\u{ffff}',
-                    '\u{ffff}', '\u{ffff}',
-                ],
-            ),
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str("OVERLAYS").unwrap(),
-                mtime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 54).unwrap(),
-                ctime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 54).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::DIRECTORY),
-                cluster: Cluster(3),
-                size: 0,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                2,
-                [
-                    '-', 'p', 'l', 'u', 's', '.', 'd', 't', 'b', '\u{0000}', '\u{ffff}',
-                    '\u{ffff}', '\u{ffff}',
-                ],
-            ),
-            Expected::Lfn(
-                false,
-                1,
-                [
-                    'b', 'c', 'm', '2', '7', '0', '8', '-', 'r', 'p', 'i', '-', 'b',
-                ],
-            ),
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str("BCM270~1.DTB").unwrap(),
-                mtime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 34).unwrap(),
-                ctime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 34).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::ARCHIVE),
-                cluster: Cluster(9),
-                size: 11120,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                1,
-                [
-                    'C', 'O', 'P', 'Y', 'I', 'N', 'G', '.', 'l', 'i', 'n', 'u', 'x',
-                ],
-            ),
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str("COPYIN~1.LIN").unwrap(),
-                mtime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 30).unwrap(),
-                ctime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 30).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::ARCHIVE),
-                cluster: Cluster(5),
-                size: 18693,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                2,
-                [
-                    'c', 'o', 'm', '\u{0}', '\u{ffff}', '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                    '\u{ffff}', '\u{ffff}', '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                ],
-            ),
-            Expected::Lfn(
-                false,
-                1,
-                [
-                    'L', 'I', 'C', 'E', 'N', 'C', 'E', '.', 'b', 'r', 'o', 'a', 'd',
-                ],
-            ),
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str("LICENC~1.BRO").unwrap(),
-                mtime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 34).unwrap(),
-                ctime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 34).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::ARCHIVE),
-                cluster: Cluster(8),
-                size: 1494,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                2,
-                [
-                    '-', 'b', '.', 'd', 't', 'b', '\u{0000}', '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                    '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                ],
-            ),
-            Expected::Lfn(
-                false,
-                1,
-                [
-                    'b', 'c', 'm', '2', '7', '0', '9', '-', 'r', 'p', 'i', '-', '2',
-                ],
-            ),
-            Expected::Short(DirEntry {
-                name: ShortFileName::create_from_str("BCM270~4.DTB").unwrap(),
-                mtime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 36).unwrap(),
-                ctime: Timestamp::from_calendar(2016, 3, 1, 19, 56, 36).unwrap(),
-                attributes: Attributes::create_from_fat(Attributes::ARCHIVE),
-                cluster: Cluster(15),
-                size: 12108,
-                entry_block: BlockIdx(0),
-                entry_offset: 0,
-            }),
-            Expected::Lfn(
-                true,
-                2,
-                [
-                    '.', 'd', 't', 'b', '\u{0000}', '\u{ffff}', '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                    '\u{ffff}', '\u{ffff}', '\u{ffff}', '\u{ffff}',
-                ],
-            ),
-            Expected::Lfn(
-                false,
-                1,
-                [
-                    'b', 'c', 'm', '2', '7', '0', '8', '-', 'r', 'p', 'i', '-', 'b',
-                ],
-            ),
-        ];
-
-        let data = parse(raw_data);
-        for (part, expected) in data.chunks(OnDiskDirEntry::LEN).zip(results.iter()) {
-            let on_disk_entry = OnDiskDirEntry::new(part);
-            match expected {
-                Expected::Lfn(start, index, contents) if on_disk_entry.is_lfn() => {
-                    let (calc_start, calc_index, calc_contents) =
-                        on_disk_entry.lfn_contents().unwrap();
-                    assert_eq!(*start, calc_start);
-                    assert_eq!(*index, calc_index);
-                    assert_eq!(*contents, calc_contents);
-                }
-                Expected::Short(expected_entry) if !on_disk_entry.is_lfn() => {
-                    let parsed_entry = on_disk_entry.get_entry(FatType::Fat32, BlockIdx(0), 0);
-                    assert_eq!(*expected_entry, parsed_entry);
-                }
-                _ => {
-                    panic!(
-                        "Bad dir entry, expected:\n{:#?}\nhad\n{:#?}",
-                        expected, on_disk_entry
-                    );
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_bpb() {
-        // Taken from a Raspberry Pi bootable SD-Card
-        const BPB_EXAMPLE: [u8; 512] = hex!(
-            "EB 3C 90 6D 6B 66 73 2E 66 61 74 00 02 10 01 00
-             02 00 02 00 00 F8 20 00 3F 00 FF 00 00 00 00 00
-             00 E0 01 00 80 01 29 BB B0 71 77 62 6F 6F 74 20
-             20 20 20 20 20 20 46 41 54 31 36 20 20 20 0E 1F
-             BE 5B 7C AC 22 C0 74 0B 56 B4 0E BB 07 00 CD 10
-             5E EB F0 32 E4 CD 16 CD 19 EB FE 54 68 69 73 20
-             69 73 20 6E 6F 74 20 61 20 62 6F 6F 74 61 62 6C
-             65 20 64 69 73 6B 2E 20 20 50 6C 65 61 73 65 20
-             69 6E 73 65 72 74 20 61 20 62 6F 6F 74 61 62 6C
-             65 20 66 6C 6F 70 70 79 20 61 6E 64 0D 0A 70 72
-             65 73 73 20 61 6E 79 20 6B 65 79 20 74 6F 20 74
-             72 79 20 61 67 61 69 6E 20 2E 2E 2E 20 0D 0A 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 00 00
-             00 00 00 00 00 00 00 00 00 00 00 00 00 00 55 AA"
-        );
-        let bpb = Bpb::create_from_bytes(&BPB_EXAMPLE).unwrap();
-        assert_eq!(bpb.footer(), Bpb::FOOTER_VALUE);
-        assert_eq!(bpb.oem_name(), b"mkfs.fat");
-        assert_eq!(bpb.bytes_per_block(), 512);
-        assert_eq!(bpb.blocks_per_cluster(), 16);
-        assert_eq!(bpb.reserved_block_count(), 1);
-        assert_eq!(bpb.num_fats(), 2);
-        assert_eq!(bpb.root_entries_count(), 512);
-        assert_eq!(bpb.total_blocks16(), 0);
-        assert_eq!(bpb.fat_size16(), 32);
-        assert_eq!(bpb.total_blocks32(), 122_880);
-        assert_eq!(bpb.footer(), 0xAA55);
-        assert_eq!(bpb.volume_label(), b"boot       ");
-        assert_eq!(bpb.fat_size(), 32);
-        assert_eq!(bpb.total_blocks(), 122_880);
-        assert_eq!(bpb.fat_type, FatType::Fat16);
-    }
-}
-
-// ****************************************************************************
-//
-// End Of File
-//
-// ****************************************************************************
